requirements.txt
django
djangorestframework
djangorestframework-simplejwt
PyJWT
python-dotenv
pillow